Improve code quality
By migrating from Psalm to PHPStan and using it, I've made the changes
in this commit. The intention is to start, regularly, modernising the
codebase, ensuring that it is as good as I can make it. One of those
ways is to keep on using static analysis, in addition to tests, and code
style checking.