Refactor the cache implementation to use PSR-16
Paraphrasing the PSR-16 documentation:

PSR-16 is a simpler approach than PSR-6 that aims to build a
standardized streamlined interface for common [caching] cases. It is
independent of PSR-6 but has been designed to make compatibility with
PSR-6 as straightforward as possible.

It's for that reason that I settled on PSR-16 instead of PSR-6. What's
more, I believe the previous implementation, receiving a class string,
was wrong. The new approach, an actual object, specifically a
CacheInterface object ensures better cache interoperability with far few
potential issues.

See https://www.php-fig.org/psr/psr-16/ for more information.