Update Composer's configuration
I don't know how to meaningfully split the configuration at this point,
bearing in mind the relationship to composer.lock. So, this change adds
psr/simple-cache (PSR-16) as a dependency, it adds PHPStan as a dev
dependency as I far prefer it to Psalm, and deprecates Psalm. It also
updates the Composer static-analysis script to use PHPStan instead of
Psalm.