Add handlers for the blog index and article view page
This change adds in default/generic handler classes providing the blog's
index page, as well as a blog article view page. In addition, naturally,
they also have accompanying templates.

The structure of the templates was the same that I used on a project,
where I have been developing the code, minus all styling, as it was very
project-specific.

I want to stress that these are generic templates and generic handlers
that are expected to be overridden at some point not long after the
package is added to a Mezzio project. But, they're there to help get
users up and running as quickly as possible.