Add laminas-inputfilter as an explicit dependency
This is because BlogArticleInputFilterFactory depends on it. I don't
know if I had it as a dependency at one stage and accidentally removed
it or whether it was never a dependency; though, I'd have thought that
it would always have had to have been one, so it was accidentally
removed. Regardless, it's back!