Add a default/fallback blog config in the ConfigProvider
The intention behind this commit is to provide a default config, should
the user not provide one, and also avoids the user having to provide
one. That way the user can get up and running more quickly, with less
time and effort.

In addition, a default config file that the can use, should they want
to, is available to be copied to config/autoload from the package's
config/autoload directory.

The README was updated to reflect these changes.

Update the README based on the revised config