Load Twig's Markdown Extension
This isn't loaded by Twig by default. However, it's used by the default
Twig templates to convert Markdown in the post files to HTML for the
content of the blog view template.

So this change adds the package and loads it into the Twig Runtime
making it always available. That said, I'm not happy about loading it if
the loading Mezzio project uses a different template engine, such as
Plates. So, in future versions, this will be, if possible, conditionally
loaded. But, for now, it is what it is.

For more information, see https://github.com/twigphp/markdown-extra.