Fix two minor bugs and static analysis issues in ConfigProvider
This small change fixes two minor bugs in setting the blog configuration
and ensuring that a value is always returned from the getTwigConfig
function. It also ensures that the function signatures are accurate and
that the docblock comments are as correct as possible to aid in complete
and thorough static analysis.