Integrate the DataTrait trait throughout the tests
This trait was set up to simplify setting up data in the tests, but
wasn't implemented throughout the ones that made sense. This change
corrects that, using it in a number of other ones, removing duplication
and potential bugs. It did require some tests to be change to reflect
the changed data being tested though.