Retrieve related articles in the BlogArticleHandler
This is something that I didn't know that I needed, I think, until
recently. That being, the ability to get a list of articles that are
related to the current one. The criteria for what makes a related
tutorial is quite simplistic: do they share the same categories or tags.
That's, honestly, it. But, for now, it works.