Fix bug in registering MarkdownToHtml with Laminas View
This change fixes a previously unseen/unknown bug with registering the
function with Laminas View's HelperPluginManager. I don't exactly know
what it is but there is some issue in how I had approached it previous
in the integration tests. What's more, the approach that I took was
unnecessary, as I found a far simpler way, demonstrated in this change.