Finish implementing pagination
This change, rough as it is, finishes implementing the final pieces of
pagination. It does the by updating all of the templates to include
next and previous links when appropriate, to include pagination when it
makes sense, and to set next and previous links if there is something to
link to in these instances.