Add a custom Twig RuntimeLoaderInterface class
This class is a custom Twig RuntimeLoaderInterface class for loading the
MarkdownRuntime. It's required so that Markdown support is available in
Twig, when required. The anonymous class implementation that I used
previously, for an as yet unknown reason, caused
data/cache/config-cache.php to break with the following error:

PHP Parse error: syntax error, unexpected token "@", expecting "]" in
/app/data/cache/config-cache.php on line 200

If you look in the file, you'll see the following at that line:

\Twig\RuntimeLoader\RuntimeLoaderInterface@anonymous^@/app/config/autoload/twig.global.php:24$d,